Refactor files to better match conventions
